noti when conflict rule

+++ expAI/views.py
@@ -126,16 +126,36 @@ class DatasetsViewSet(viewsets.ModelViewSet):
     # def perform_create(self, serializer):
     #     serializer.save(datasetowner=self.request.user)
 
+    def update(self, request, *args, **kwargs):
+        partial = kwargs.pop('partial', False)
+        usr = self.request.user
+        instance = self.get_object()
+        if usr.roleid.rolename != "ADMIN" or int(usr.id) != int(instance.datasetowner):
+            return Response("Tài khoản cần Admin hoặc sở hữu dataset này",status=status.HTTP_401_UNAUTHORIZED)
+        serializer = self.get_serializer(instance, data=request.data, partial=partial)
+        serializer.is_valid(raise_exception=True)
+        self.perform_update(serializer)
+
+        if getattr(instance, '_prefetched_objects_cache', None):
+            # If 'prefetch_related' has been applied to a queryset, we need to
+            # forcibly invalidate the prefetch cache on the instance.
+            instance._prefetched_objects_cache = {}
+
+        return Response(serializer.data)
     def destroy(self, request, *args, **kwargs):
+        # try:
+        usr = self.request.user
+        print(usr.roleid.rolename)
+        instance = self.get_object()
+        if usr.roleid.rolename != "ADMIN" and int(usr.id) != int(instance.datasetowner.id):
+            return Response("Tài khoản cần Admin hoặc sở hữu dataset này",status=status.HTTP_401_UNAUTHORIZED)
         try:
-            instance = self.get_object()
-            try:
-                import shutil
-                shutil.rmtree(f'datasets/{instance.datasetfolderurl}')
-            except:...
-            self.perform_destroy(instance)
-        except:
-            return Response(status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            import shutil
+            shutil.rmtree(f'datasets/{instance.datasetfolderurl}')
+        except:...
+        self.perform_destroy(instance)
+        # except:
+        #     return Response(status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
         return Response(status=status.HTTP_204_NO_CONTENT)
